Revert "fix: fix bugs and add css to edit, delete and submit buttons"
This reverts commit feb18f41c91110ea43179765ae1037275df08087.
