+++ data/db.json
@@ -95,6 +95,22 @@
       "datashared": false,
       "peerreview": true,
       "enforced": "Yes - before application"
+    },
+    {
+      "id": 7,
+      "title": "TCP/IP Architecture, Design, and Implementation in Linux",
+      "authors": "Sm Zobaed, Sameer Seth",
+      "journaltype": "Book",
+      "topic": "Networking / LINUX",
+      "published": "29 January 2008",
+      "issn": "issn-615-892643",
+      "updated": "16th Aug 2022",
+      "link": "https://doi.org/10.1002/agg2.20287",
+      "policy": "Policy 1",
+      "dataavail": true,
+      "datashared": false,
+      "peerreview": true,
+      "enforced": "Yes - before application"
     }
   ]
 }
