+++ src/components/JournalDetails/Details.js
@@ -1,4 +1,3 @@
-/* eslint-disable jsx-a11y/anchor-is-valid */
 /* eslint-disable react/button-has-type */
 /* eslint-disable no-shadow */
 /* eslint-disable react/jsx-key */
@@ -28,10 +27,8 @@ import {
   Misc,
   UpdateContainer,
   Icon,
-  ButtonContainer,
 } from './styles';
 import { Authors, Head3 } from '../Journals/styles';
-import { FormInputBtn } from '../Authentication/styles';
 
 function Details({ posts, handleDelete }) {
   const { id } = useParams();
@@ -129,14 +126,10 @@ function Details({ posts, handleDelete }) {
                 ))}
               </UpdateContainer>
             </Misc>
-            <ButtonContainer>
-              <Link to={`/edit/${indv.id}`}>
-                <FormInputBtn>Edit Post</FormInputBtn>
-              </Link>
-              <FormInputBtn style={{ marginLeft: '1rem' }} onClick={() => handleDelete(indv.id)}>
-                Delete Post
-              </FormInputBtn>
-            </ButtonContainer>
+            <Link to={`/edit/${indv.id}`}>
+              <button>Edit Post</button>
+            </Link>
+            <button onClick={() => handleDelete(indv.id)}>Delete Post</button>
           </div>
         </PolicyContainer>
       )}
