+++ src/components/marginals/Navbar/Navbar.js
@@ -15,6 +15,7 @@ function Navbar({ search, setSearch }) {
     <Nav>
       <Logo src={navbar.logo.src} />
       <form className='searchForm' onSubmit={(e) => e.preventDefault()}>
+        <label htmlFor='search'>Search Journal</label>
         <input
           id='search'
           type='text'
