+++ src/components/marginals/Search/Search.js
@@ -7,14 +7,29 @@ import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 import '../../../index.css';
 
-function SearchBar({
-  posts,
-  filteredData,
-  setFilteredData,
-  wordEntered,
-  setWordEntered,
-  handleFilter,
-}) {
+function SearchBar({ posts }) {
+  const [filteredData, setFilteredData] = useState([]);
+  const [wordEntered, setWordEntered] = useState('');
+
+  const handleFilter = (event) => {
+    const searchWord = event.target.value;
+    setWordEntered(searchWord);
+    const newFilter = posts.filter((value) => {
+      return value.title.toLowerCase().includes(searchWord.toLowerCase());
+    });
+
+    if (searchWord === '') {
+      setFilteredData([]);
+    } else {
+      setFilteredData(newFilter);
+    }
+  };
+
+  const clearInput = () => {
+    setFilteredData([]);
+    setWordEntered('');
+  };
+
   return (
     <div className='search'>
       <div className='searchInputs'>
