+++ src/index.css
@@ -82,6 +82,10 @@ a {
   text-decoration: none;
 }
 
+/* a:hover {
+  background-color: lightgrey;
+} */
+
 #clearBtn {
   cursor: pointer;
 }
