+++ src/pages/Home.js
@@ -2,30 +2,10 @@ import React from 'react';
 import { Header, LandingSection } from '../components';
 import '../components/Landing/Header/Header.css';
 
-function Home({
-  search,
-  setSearch,
-  searchResults,
-  posts,
-  filteredData,
-  setFilteredData,
-  wordEntered,
-  setWordEntered,
-  handleFilter,
-}) {
+function Home({ search, setSearch, searchResults, posts }) {
   return (
     <div className='header-padding'>
-      <Header
-        search={search}
-        setSearch={setSearch}
-        searchResults={searchResults}
-        posts={posts}
-        filteredData={filteredData}
-        setFilteredData={setFilteredData}
-        wordEntered={wordEntered}
-        setWordEntered={setWordEntered}
-        handleFilter={handleFilter}
-      />
+      <Header search={search} setSearch={setSearch} searchResults={searchResults} posts={posts} />
       <LandingSection />
     </div>
   );
